extend filter renamed to merge
 #805
