+++ waltz-ng/client/common/filters/merge-filter.js
