+++ waltz-ng/client/playpen/2/playpen2.html
@@ -12,7 +12,7 @@
     <waltz-section name="Search" small="All">
 
 
-        <waltz-entity-link entity-ref="ctrl.entityReference | extend:({kind: 'ORG_UNIT'})"
+        <waltz-entity-link entity-ref="ctrl.entityReference | merge:({kind: 'ORG_UNIT'})"
                            icon-placement="'right'"
                            tooltip-placement="'right'">
         </waltz-entity-link>
